// database.types.ts
export type Database = {
    public: {
        Tables: {
            equipment: {
                Row: {
                    avg_consumption_per_washing_cycle: number | null
                    avg_monthly_consumption_fan: number | null
                    brand: string | null
                    created_at: string
                    daily_standby_consumption_kwh: number | null
                    daily_usage_hours: number | null
                    hourly_consumption_in_use_kwh: number | null
                    id: number
                    location: number
                    model: string | null
                    monthly_consumption_kwh: number | null
                    name: string
                    power: number | null
                    user_id: string
                }
                Insert: {
                    avg_consumption_per_washing_cycle?: number | null
                    avg_monthly_consumption_fan?: number | null
                    brand?: string | null
                    created_at?: string
                    daily_standby_consumption_kwh?: number | null
                    daily_usage_hours?: number | null
                    hourly_consumption_in_use_kwh?: number | null
                    id?: number
                    location: number
                    model?: string | null
                    monthly_consumption_kwh?: number | null
                    name: string
                    power?: number | null
                    user_id?: string
                }
                Update: {
                    avg_consumption_per_washing_cycle?: number | null
                    avg_monthly_consumption_fan?: number | null
                    brand?: string | null
                    created_at?: string
                    daily_standby_consumption_kwh?: number | null
                    daily_usage_hours?: number | null
                    hourly_consumption_in_use_kwh?: number | null
                    id?: number
                    location?: number
                    model?: string | null
                    monthly_consumption_kwh?: number | null
                    name?: string
                    power?: number | null
                    user_id?: string
                }
                Relationships: [
                    {
                        foreignKeyName: "equipment_location_fkey"
                        columns: ["location"]
                        isOneToOne: false
                        referencedRelation: "location"
                        referencedColumns: ["id"]
                    },
                ]
            }
            location: {
                Row: {
                    created_at: string
                    id: number
                    name: string
                    user_id: string
                }
                Insert: {
                    created_at?: string
                    id?: number
                    name: string
                    user_id: string
                }
                Update: {
                    created_at?: string
                    id?: number
                    name?: string
                    user_id?: string
                }
                Relationships: []
            }
        }
        Views: {
            [_ in never]: never
        }
        Functions: {
            [_ in never]: never
        }
        Enums: {
            [_ in never]: never
        }
        CompositeTypes: {
            [_ in never]: never
        }
    }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
    DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
    EnumName extends DefaultSchemaEnumNameOrOptions extends {
        schema: keyof Database
    }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
    ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
    : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
    PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof Database },
    CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
        schema: keyof Database
    }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
    ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
    : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never